+++ go/fennel/model/counter/list_test.go
@@ -156,7 +156,7 @@ func TestList_Bucketize_Valid(t *testing.T) {
 		actions.Append(d)
 		expected = append(expected, counter.Bucket{Window: ftypes.Window_DAY, Index: 1, Width: 1, Key: v.String()})
 		expVals = append(expVals, value.NewList(e))
-		expected = append(expected, counter.Bucket{Key: v.String(), Window: ftypes.Window_MINUTE, Index: uint64(24*10 + i*10), Width: 6})
+		expected = append(expected, counter.Bucket{Key: v.String(), Window: ftypes.Window_MINUTE, Index: uint32(24*10 + i*10), Width: 6})
 		expVals = append(expVals, value.NewList(e))
 	}
 	buckets, vals, err := Bucketize(h, actions)
