+++ go/fennel/model/counter/storage3_test.go
@@ -41,7 +41,7 @@ func TestThirdStore(t *testing.T) {
 				Key:    "key",
 				Window: ftypes.Window_FOREVER,
 				Width:  120,
-				Index:  uint64(j),
+				Index:  uint32(j),
 			}
 			vals[i][j] = value.Int(i * j)
 		}
